Fix a similar size_t/unsigned int mismatch.

+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3model.cpp
@@ -37,10 +37,10 @@ vector<Mesh*> meshes(0);
 
 namespace enigma {
 
-unsigned int split(const std::string &txt, std::vector<std::string> &strs, char ch)
+size_t split(const std::string &txt, std::vector<std::string> &strs, char ch)
 {
-	unsigned int pos = txt.find( ch );
-	unsigned int initialPos = 0;
+	size_t pos = txt.find( ch );
+	size_t initialPos = 0;
 	strs.clear();
 
 	// Decompose statement
@@ -52,7 +52,7 @@ unsigned int split(const std::string &txt, std::vector<std::string> &strs, char 
 	}
 
 	// Add the last one
-	strs.push_back( txt.substr( initialPos, std::min( pos, (unsigned)txt.size() ) - initialPos + 1 ) );
+	strs.push_back( txt.substr( initialPos, std::min( pos, txt.size() ) - initialPos + 1 ) );
 
 	return strs.size();
 }
